bcctrx opcode will fall back to the interpreter, and a minor compile and warning fix.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3454 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/VideoInterface.cpp
@@ -1050,7 +1050,7 @@ void Update()
 	// Calculate actual refresh rate
 	static u64 LastTick = 0;
 	static int UpdateCheck = timeGetTime() + 1000, TickProgress = 0;
-	if (UpdateCheck < timeGetTime())
+	if (UpdateCheck < (int)timeGetTime())
 	{
 		UpdateCheck = timeGetTime() + 1000;
 		TickProgress = CoreTiming::GetTicks() - LastTick;
